# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
...
      "AltaDeNegocio": {
        "properties": {
          "categoria": {
            "description": "Slug de una categoría global, por ejemplo «barberia»",
            "title": "Categoria",
            "type": "string"
          },
          "direccion": {
            "title": "Direccion",
            "type": "string"
          },
          "latitud": {
            "title": "Latitud",
            "type": "number"
          },
          "longitud": {
            "title": "Longitud",
            "type": "number"
          },
          "nombre": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Nombre",
            "type": "string"
          },
          "zona_horaria": {
            "default": "America/Panama",
            "title": "Zona Horaria",
            "type": "string"
          }
        },
        "required": [
          "nombre",
          "categoria",
          "direccion",
          "longitud",
          "latitud"
        ],
        "title": "AltaDeNegocio",
        "type": "object"
      },
...
      "NegocioCreado": {
        "properties": {
          "estado": {
            "title": "Estado",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "estado"
        ],
        "title": "NegocioCreado",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Reservas y marketplace de belleza y bienestar en Panamá. Cada endpoint cita el requisito del brief que cubre.",
    "title": "API de Bukeo",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/negocios": {
      "post": {
        "description": "Crea el negocio, la membresía de dueño y sus ajustes. **Sin tarjeta y sin esperar a nadie.**\n\nNace en `borrador`: solo se publica cuando cumple el mínimo (D11), y publicar es un paso\naparte y explícito. Un negocio a medias en el marketplace es peor que uno que no está.",
        "operationId": "crear_negocio_api_v1_negocios_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AltaDeNegocio"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NegocioCreado"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alta self-service de un negocio (ONB-2)",
        "tags": [
          "alta de negocio"
        ]
      }
    },
    "/api/v1/publico/buscar": {
      "get": {
        "description": "Busca por texto, categoría, zona o cercanía, filtra y ordena.\n\nBuscar «por zona» no es lo mismo que buscar «cerca de mí»: la zona es un sitio con nombre\nque la gente escribe en Google, y el radio es el gesto de quien tiene el GPS encendido. Se\npuede combinar, y quien pide «Bella Vista» ve también El Cangrejo y Obarrio, que están\ndentro.\n\n**El filtro de disponibilidad usa el mismo motor que la reserva**, no una copia (MKT-2):\nprometer horas libres con una fórmula distinta a la de reservar es mandar a la gente a un\nhueco que no existe. Es también el filtro más caro, así que se aplica al final y solo sobre\nlos candidatos que ya pasaron todo lo demás.\n\nLos patrocinados se intercalan aquí cuando existan (Fase 4): **se insertan entre los\norgánicos, nunca en su lugar**, y siempre etiquetados.",
        "operationId": "buscar_api_v1_publico_buscar_get",
        "parameters": [
          {
            "description": "Nombre del negocio o del servicio",
            "in": "query",
            "name": "texto",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Nombre del negocio o del servicio",
              "title": "Texto"
            }
          },
          {
            "in": "query",
            "name": "categoria",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Categoria"
            }
          },
          {
            "description": "Slug de zona, por ejemplo «el-cangrejo»",
            "in": "query",
            "name": "zona",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Slug de zona, por ejemplo «el-cangrejo»",
              "title": "Zona"
            }
          },
          {
            "in": "query",
            "name": "longitud",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longitud"
            }
          },
          {
            "in": "query",
            "name": "latitud",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Latitud"
            }
          },
          {
            "in": "query",
            "name": "radio_metros",
            "required": false,
            "schema": {
              "default": 10000,
              "maximum": 50000,
              "minimum": 200,
              "title": "Radio Metros",
              "type": "integer"
            }
          },
          {
            "description": "En centavos. Filtra por ALGÚN servicio en rango",
            "in": "query",
            "name": "precio_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "En centavos. Filtra por ALGÚN servicio en rango",
              "title": "Precio Min"
            }
          },
          {
            "description": "En centavos",
            "in": "query",
            "name": "precio_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "En centavos",
              "title": "Precio Max"
            }
          },
          {
            "description": "Sobre el rating bayesiano, no la media",
            "in": "query",
            "name": "rating_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 5,
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sobre el rating bayesiano, no la media",
              "title": "Rating Min"
            }
          },
          {
            "description": "Disponibilidad real, calculada con el motor de reservas",
            "in": "query",
            "name": "disponibilidad",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(ahora|hoy|fecha)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Disponibilidad real, calculada con el motor de reservas",
              "title": "Disponibilidad"
            }
          },
          {
            "description": "Obligatorio con disponibilidad=fecha",
            "in": "query",
            "name": "dia",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Obligatorio con disponibilidad=fecha",
              "title": "Dia"
            }
          },
          {
            "in": "query",
            "name": "abierto_ahora",
            "required": false,
            "schema": {
              "default": false,
              "title": "Abierto Ahora",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "orden",
            "required": false,
            "schema": {
              "default": "relevancia",
              "pattern": "^(relevancia|distancia|precio|rating|nuevos)$",
              "title": "Orden",
              "type": "string"
            }
          },
          {
            "description": "Calcula la primera hora libre de hoy de cada resultado. Cuesta una consulta de agenda por negocio: se pide cuando la pantalla lo va a pintar",
            "in": "query",
            "name": "con_proxima_hora",
            "required": false,
            "schema": {
              "default": false,
              "description": "Calcula la primera hora libre de hoy de cada resultado. Cuesta una consulta de agenda por negocio: se pide cuando la pantalla lo va a pintar",
              "title": "Con Proxima Hora",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "pagina",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Pagina",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResultadoDeBusqueda"
                  },
                  "title": "Response Buscar Api V1 Publico Buscar Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Búsqueda del marketplace (MKT-1, MKT-2, MKT-3, MKT-4)",
        "tags": [
          "público"
        ]
      }
    },
...
    "/api/v1/publico/negocios/{slug}": {
      "get": {
        "description": "El perfil que indexa Google y desde el que se reserva.\n\nLleva servicios con precio y duración, equipo visible, fotos, horario, atributos y el\nrating agregado. **No lleva el teléfono**: el click-to-chat se resuelve en servidor, porque\nsi el número viajara aquí, alguien se lleva la base entera de negocios en una tarde.\n\nTodo sale con el rol del marketplace, que **no tiene permiso** sobre reservas ni sobre\nfichas de cliente: aunque este código quisiera, no podría filtrar un dato de una persona.",
        "operationId": "perfil_api_v1_publico_negocios__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilPublico"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Perfil público del negocio (NEG-1, NEG-3)",
        "tags": [
          "público"
        ]
      }
    },
    "/api/v1/publico/negocios/{slug}/chat": {
      "get": {
        "description": "Registra el clic y redirige. **El número nunca llega al navegador.**\n\nEs la garantía nº 3 del proyecto hecha código: si el teléfono viajara en el perfil, un\nscript se lleva la base entera de negocios de Panamá en una tarde. Aquí el cliente pide una\nURL de nuestro dominio, el servidor apunta el clic y responde con la redirección; el número\nsolo existe entre la base de datos y esta función.\n\nEl clic se cuenta **agregado por día** (MKT-8): 5.000 negocios generando una fila por\nevento no aporta nada que la serie no cuente mejor.",
        "operationId": "click_to_chat_api_v1_publico_negocios__slug__chat_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "307": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Salto a WhatsApp resuelto en servidor (NEG-1, MKT-8)",
        "tags": [
          "público"
        ]
      }
    },
...
  }
}